# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BookWorm API
  description: |
    Full-stack online bookstore REST API.
    
    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    
    ```
    Authorization: Bearer <token>
    ```
    
    ## Getting Started
    1. Register at `POST /auth/register` or login at `POST /auth/login`
    2. Copy the returned `token`
    3. Click "Authorize" above and paste: `Bearer <token>`
  version: 1.0.0
  contact:
    name: BookWorm API Support
servers:
  - url: http://localhost:5000/api/v1
    description: Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
        subtotal:
          type: number
        tax_amount:
          type: number
        discount_amount:
          type: number
        delivery_charges:
          type: number
        grand_total:
          type: number
        order_status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        payment_status:
          type: string
          enum: [pending, completed, failed, refunded]
        estimated_delivery_date:
          type: string
          format: date
        can_cancel:
          type: boolean
...
security:
  - bearerAuth: []
paths:
...
  /cart/items:
    post:
      tags: [Cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [book_id, quantity]
              properties:
                book_id:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 10
      responses:
        '201':
          description: Item added
  /cart/items/{cartItemId}:
    put:
      tags: [Cart]
      summary: Update cart item quantity
      parameters:
        - in: path
          name: cartItemId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Cart item updated
    delete:
      tags: [Cart]
      summary: Remove item from cart
      parameters:
        - in: path
          name: cartItemId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Item removed

  # ── Addresses ───────────────────────────────────────────────────────
...
  /orders/create:
    post:
      tags: [Orders]
      summary: Create order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address_id]
              properties:
                address_id:
                  type: string
                  format: uuid
                gift_points_to_redeem:
                  type: integer
                  minimum: 0
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order details
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details with items
  /orders/{orderId}/cancel:
    put:
      tags: [Orders]
      summary: Cancel order (within 48 hours)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order cancelled
        '400':
          description: Cannot cancel after 48 hours
  /orders/{orderId}/buy-again:
    post:
      tags: [Orders]
      summary: Add past order items to cart
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Items added to cart
...